Streamline fedmsg diff publication.
We used to publish a diff of the whole repo.. but in some cases, that
would produce gigantic messages (I found a few that were 25MB each).
I'm almost certain that this is what is causing datagrepper to choke
under OOM.

This change restructures the messages, simplifying them, so that we do
not publish a whole diff, but only a list of the packages that had some
change.

This requires two other changes, one to fedora-packages, which uses
these messages to update itself.  And another to fedmsg.meta which
is used to produce human-readable strings about these messages.

- https://github.com/fedora-infra/fedora-packages/pull/225
- https://github.com/fedora-infra/fedmsg_meta_fedora_infrastructure/pull/358