Update how we run the application to latest aiohttp.
This commit updates how we run mdapi, aiohttp makes it easier to
start the application without having to deal with the event loop.
In this commit we also create a dedicated server module which job
is to start the application.

We also make use of the request query property instead of trying to
parse the query string.

Signed-off-by: Clement Verna <cverna@tutanota.com>