Facilitate containerized deployment
Virtual environment is not required within the container image
and hence the configuration is set to false to ensure that
Poetry does not end up creating a virtual environment by itself
in the absence of an existing one.

Also, as `setuptools` is not previously mentioned in the
dependency list of `pyproject.toml`, and that is why during
installing the project, Poetry uninstalls `setuptools`. Hence,
this necessitates explicitly specifying `setuptools` in the
`pyproject.toml` dependency list. Thanks, @abompard for marking
that out.

Signed-off-by: Akashdeep Dhar <akashdeep.dhar@gmail.com>