Report correct version for SRPM with multiple subpackages
If there is multiple subpackages from single SRPM, the first subpackage
is taken. This is fine as long as the subpackages are reusing the main
package versioning, which might not always be the case.

To avoid issue, from the list of subpackages, try to select first the
package which matches the source package name.

Unfortunately, this still leaves room for issues, where there is no
subpackage of the same name as SRPM and when the versioning of
subpackages differs from the SRPM version.

Fixes #35