Avoid skipping valid branches while indexing repos
It looks like #296 introduced a subtle bug.  Previously the
list_branches function returned more branches than were needed (despite
the clt_status param), so we needed to skip invalid branches.  The skip
condition accidentally always evaluated to true, so valid branches were
being skipped too.  With the switch to querying bodhi for current
branches, we only get valid branches.  That makes it so that the easiest
fix is to just remove the skip condition.

Signed-off-by: Carl George <carlwgeorge@gmail.com>